Makefile.in: Fix terrible omission and handle broken LDFLAGS better.
The terrible omission was a lack of comma definition, leading to wrong
options passed to $(CCLD) when LDFLAGS is non-empty.

Broken LDFLAGS is the one assuming that LD=$(CC), thus implicitly adding
-Wl, in front of all linker's options (or one -Wl, in front and commas
instead of spaces). Sadly this is a common assumption, still a bad one.

Now vidma-git in AUR (Arch User Repository) should be buildable.

NOTE: In future Makefile.in will be updated to be closer to my latest
      "template" that I wrote for pecutils.